Add Update Event Task table and move GitHub Summary feature (#158)
