--- busy_beaver/tasks/__init__.py
@@ -1,1 +1,0 @@
-from .github_stats import start_post_github_summary_task  # noqa
